requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
numpy==1.23.5
opencv-python==4.8.1.78
psycopg2-binary==2.9.6
onnxruntime==1.22.0
insightface==0.7.3
pgvector==0.2.4